+++ apps/mobile-app/CHANGELOG.md
@@ -1,5 +1,12 @@
 # bestaff-mobile-app
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @bestaff/ui@0.0.2
+
 ## 1.0.1
 
 ### Patch Changes